// src/api/generated/models/taskDtoState.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * TodoList API
 * API para gestión de tareas y usuarios
 * OpenAPI spec version: 1.0
 */

/**
 * Estado actual de la tarea
 */
export type TaskDtoState = (typeof TaskDtoState)[keyof typeof TaskDtoState];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TaskDtoState = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;